Split out OS test utils into new folder
