+++ CMakeLists.txt
@@ -25,7 +25,7 @@ include(CodeCoverage)
 if(NOT DEFINED REALM_VERSION)
     # Get accurate git-describe version
     include(GetGitRevisionDescription)
-    git_describe(REALM_VERSION)
+    #git_describe(REALM_VERSION)
 endif()
 set(PROJECT_VERSION ${DEP_VERSION})
