+++ test/object-store/CMakeLists.txt
@@ -42,15 +42,16 @@ file(GLOB RESOURCES RELATIVE ${CMAKE_CURRENT_BINARY_DIR}
      "*.realm")
 
 if(REALM_ENABLE_AUTH_TESTS)
-    list(APPEND SOURCES util/baas_admin_api.cpp)
+    list(APPEND SOURCES util/baas/baas_admin_api.cpp)
 endif()
 
 if(REALM_ENABLE_SYNC)
     list(APPEND HEADERS
         sync/flx_sync_harness.hpp
-        sync/session/session_util.hpp
-        sync/sync_test_utils.hpp
-        util/baas_admin_api.hpp
+        util/session/session_util.hpp
+        util/sync/sync_test_utils.hpp
+        util/sync/common_utils.hpp
+        util/baas/baas_admin_api.hpp
     )
     list(APPEND SOURCES
         bson.cpp
@@ -64,7 +65,8 @@ if(REALM_ENABLE_SYNC)
         sync/session/session.cpp
         sync/session/wait_for_completion.cpp
         sync/sync_manager.cpp
-        sync/sync_test_utils.cpp
+        util/sync/common_utils.cpp
+        util/sync/sync_test_utils.cpp
         sync/user.cpp
         sync/app.cpp
     )
