+++ test/object-store/audit.cpp
@@ -22,7 +22,8 @@
 #include "util/event_loop.hpp"
 #include "util/test_file.hpp"
 #include "util/test_utils.hpp"
-#include "util/baas_admin_api.hpp"
+#include "util/sync/sync_test_utils.hpp"
+#include "util/baas/baas_admin_api.hpp"
 
 #include <realm/object-store/audit.hpp>
 #include <realm/object-store/audit_serializer.hpp>
