+++ test/object-store/sync/flx_sync_harness.hpp
@@ -21,8 +21,8 @@
 
 #ifdef REALM_ENABLE_AUTH_TESTS
 
-#include "sync/sync_test_utils.hpp"
-#include "util/baas_admin_api.hpp"
+#include "../util/baas/baas_admin_api.hpp"
+#include "../util/sync/sync_test_utils.hpp"
 
 namespace realm::app {
