+++ test/object-store/sync/session/wait_for_completion.cpp
@@ -17,9 +17,9 @@
 ////////////////////////////////////////////////////////////////////////////
 
 #include <catch2/catch_all.hpp>
-#include "sync/session/session_util.hpp"
-#include "util/event_loop.hpp"
-#include "util/test_utils.hpp"
+#include "../../util/session/session_util.hpp"
+#include "../../util/event_loop.hpp"
+#include "../../util/test_utils.hpp"
 
 #include <realm/util/scope_exit.hpp>
