+++ test/object-store/sync/sync_manager.cpp
@@ -16,14 +16,14 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-#include "sync_test_utils.hpp"
+#include "../util/sync/sync_test_utils.hpp"
 
 #include <realm/sync/config.hpp>
 #include <realm/object-store/sync/sync_manager.hpp>
 #include <realm/object-store/sync/sync_user.hpp>
-#include "sync/session/session_util.hpp"
-#include "util/event_loop.hpp"
-#include "util/test_utils.hpp"
+#include "../util/session/session_util.hpp"
+#include "../util/event_loop.hpp"
+#include "../util/test_utils.hpp"
 
 #include <realm/util/logger.hpp>
 #include <realm/util/optional.hpp>
