+++ test/object-store/util/test_catch_macros.hpp
@@ -16,43 +16,11 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-#ifndef REALM_TEST_UTILS_HPP
-#define REALM_TEST_UTILS_HPP
+#ifndef REALM_TEST_CATCH_MACROS_HPP
+#define REALM_TEST_CATCH_MACROS_HPP
 
 #include <catch2/catch_all.hpp>
 #include <catch2/matchers/catch_matchers_all.hpp>
-#include <realm/util/file.hpp>
-#include <realm/util/optional.hpp>
-#include "sync/common_utils.hpp"
-
-#include <functional>
-#include <filesystem>
-namespace fs = std::filesystem;
-
-namespace realm {
-
-/// Open a Realm at a given path, creating its files.
-bool create_dummy_realm(std::string path);
-void reset_test_directory(const std::string& base_path);
-std::vector<char> make_test_encryption_key(const char start = 0);
-void catch2_ensure_section_run_workaround(bool did_run_a_section, std::string section_name,
-                                          util::FunctionRef<void()> func);
-
-std::string encode_fake_jwt(const std::string& in, util::Optional<int64_t> exp = {},
-                            util::Optional<int64_t> iat = {});
-
-static inline int64_t random_int()
-{
-    thread_local std::mt19937_64 rng(std::random_device{}());
-    return rng();
-}
-
-bool chmod_supported(const std::string& path);
-int get_permissions(const std::string& path);
-void chmod(const std::string& path, int permissions);
-std::string get_parent_directory(const std::string& path);
-
-} // namespace realm
 
 #define REQUIRE_DIR_EXISTS(macro_path)                                                                               \
     do {                                                                                                             \
@@ -90,6 +58,4 @@ std::string get_parent_directory(const std::string& path);
 
 #define REQUIRE_THROWS_CONTAINING(expr, msg) REQUIRE_THROWS_WITH(expr, Catch::Matchers::ContainsSubstring(msg))
 
-#define ENCODE_FAKE_JWT(in) realm::encode_fake_jwt(in)
-
-#endif // REALM_TEST_UTILS_HPP
+#endif // REALM_TEST_CATCH_MACROS_HPP
